refactor(rentman): reine Mapping-Helfer aus RentmanImportDialog auslagern (#468)
RentmanImportDialog.tsx 2140 → 1784 Zeilen. Die ~360 Zeilen reiner Daten-/
Parsing-Helfer (kein React/JSX) wandern nach rentmanImportHelpers.ts:
mapProjects, mapEquipment, detectCableRows, autoDetectCategory, die
Kategorie-Map-Persistenz u. a. + die zugehörigen Typen.

Verhaltensneutral: identischer Code, nur verschoben + importiert. Macht die
Parser einzeln testbar und den Dialog überschaubarer. tsc 0, eslint 0,
build 0.

Refs #468
